RoadManager: I forgot to convert the last bit into an int. Now it's actually, completely fixed. Thanks!

+++ Assets/scripts/RoadManager.cs
@@ -110,7 +110,7 @@ public class RoadManager : MonoBehaviour
 
         segments = new Segment[(int)(trackLength/segmentLength)];
 
-        segmentToCalculateLoopAt = trackLength - 50;
+        segmentToCalculateLoopAt = (int)(trackLength - 50);
     }
 
     // Start is called before the first frame update
